STL-1821 | API v1 restructured, leaned to FastAPI
