+++ erica/api/v1/endpoints/grundsteuer/grundsteuer.py
@@ -1,0 +1,20 @@
+from fastapi import status, APIRouter
+from erica.request_processing.erica_input import GrundsteuerData
+
+router = APIRouter()
+
+
+@router.post('/', status_code=status.HTTP_201_CREATED)
+def send_grundsteuer(grundsteuer_data: GrundsteuerData, include_elster_responses: bool = False):
+    """
+    The Grundsteuer data is validated and then send to ELSTER using ERiC. If it is successful, this should return a 201
+    HTTP response with {'transfer_ticket': str, 'pdf': str}. The pdf is base64 encoded binary data of the pdf
+    If there is any error with the validation, this should return a 400 response. If the validation failed with
+    {‘code’ : int,‘message’: str,‘description’: str, ‘validation_problems’ : [{‘code’: int, ‘message’: str}]}
+    or a 400 repsonse for other client errors and a 500 response for server errors with
+    {‘code’ : int,‘message’: str,‘description’: str}
+
+    :param grundsteuer_data: the JSON input data for the land tax declaration
+    :param include_elster_responses: query parameter which indicates whether the ERiC/Server response are returned
+    """
+    return {"request": "successful"}
