+++ erica/api/v2/api_v2.py
@@ -1,0 +1,4 @@
+from fastapi import APIRouter
+
+api_router_02 = APIRouter()
+api_router_02.prefix = '/02'
