+++ erica/api/v_none/ping.py
@@ -1,0 +1,10 @@
+from fastapi import APIRouter
+
+router = APIRouter()
+
+
+@router.get('/')
+def ping():
+    """Simple route that can be used to check if the app has started.
+    """
+    return 'pong'
